+++ packages/patrol_cli/lib/src/commands/test.dart
@@ -320,6 +320,10 @@ See https://github.com/leancodepl/patrol/issues/1316 to learn more.
       serverTimeout: intArg('web-server-timeout'),
       browserArgs: stringArg('web-browser-args'),
       baseUrl: stringArg('web-base-url'),
+      optimizationLevel: stringArg('optimization-level'),
+      sourceMaps: argResults!.wasParsed('source-maps')
+          ? boolArg('source-maps')
+          : null,
     );
 
     // No need to build web app for testing. It's done in the execute method.